backend/requirements.txt
fastapi
uvicorn
pypdf
python-docx
openpyxl
python-pptx
sentence-transformers
faiss-cpu
requests
python-dotenv
python-multipart
